Modularize & deprecate simulation builder
