+++ openfisca_core/simulations/__init__.py
@@ -23,5 +23,6 @@
 
 from openfisca_core.errors import CycleError, NaNCreationError, SpiralError  # noqa: F401
 
-from .helpers import calculate_output_add, calculate_output_divide  # noqa: F401
+from .helpers import calculate_output_add, calculate_output_divide, check_type, transform_to_strict_syntax  # noqa: F401
 from .simulation import Simulation  # noqa: F401
+from .simulation_builder import SimulationBuilder  # noqa: F401
